requirement.txt
openai
llama-index
llama-index-embeddings-clip
git+https://github.com/openai/CLIP.git
kaggle
streamlit-carousel
qdrant_client